major modifications

+++ Framework/IP-Framework/API/Response.cs
@@ -7,7 +7,7 @@ namespace IP_Framework.API
     public class Response
     {
         int id { get; set; }
-        int status { get; set; }
+        double status { get; set; }
 
     }
 }
